PUSH

+++ game/backgroundhandler.c
@@ -5,7 +5,7 @@ Background menuBackground;
 Background skyBackground;
 
 // Function prototypes
-void initializeBackground(Background *background, int *asset, int width, int height) {
+void initializeBackground(Background *background, short int *asset, int width, int height) {
     background->cols = (WORLD_WIDTH / width) + (WORLD_WIDTH % width != 0); // Ceiling division
     background->rows = (WORLD_HEIGHT / height) + (WORLD_HEIGHT % height != 0); // Ceiling division
